requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.0
requests>=2.31.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
trafilatura>=1.9.0

# Docs
mkdocs>=1.6.0
mkdocs-material>=9.5.10
mkdocs-include-markdown-plugin>=6.0.0
pymdown-extensions>=10.8

# Optional
feedparser>=6.0.10
sqlite-utils>=3.36

# Job processing
redis>=5.0.0
apscheduler>=3.10.0

# Embeddings for semantic search
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
